refactor service layer & endpoint layer

+++ server/src/main/java/ru/trandefil/tm/endpoint/UserEndPointImpl.java
@@ -4,6 +4,7 @@ import ru.trandefil.tm.api.UserService;
 import ru.trandefil.tm.entity.Session;
 import ru.trandefil.tm.entity.User;
 import ru.trandefil.tm.generated.UserEndPoint;
+import ru.trandefil.tm.util.SignatureUtil;
 
 import javax.jws.WebMethod;
 import javax.jws.WebService;
@@ -21,30 +22,50 @@ public class UserEndPointImpl implements UserEndPoint {
     @WebMethod
     @Override
     public User deleteUser(User user, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return userService.delete(user,session);
     }
 
     @WebMethod
     @Override
     public User deleteUserByName(String name,Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return userService.deleteByName(name,session);
     }
 
     @WebMethod
     @Override
     public User saveUser(User user,Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return userService.save(user,session);
     }
 
     @WebMethod
     @Override
     public User getUserByName(String userName,Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return userService.getByName(userName,session);
     }
 
     @WebMethod
     @Override
     public List<User> getAllUsers(Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return userService.getAll(session);
     }
